backend/requirements.txt
Django>=3.0,<4.0
gunicorn>=20.0
psycopg2-binary>=2.8
requests
iso8601
python-dotenv
djangorestframework
django-cors-headers
djangorestframework-simplejwt